# client-web/src/types/ticket.ts
import type { components } from './api.generated'

export type Priority = components['schemas']['Priority']
export type TicketStatus = components['schemas']['TicketStatus']
export type AssignmentMode = components['schemas']['AssignmentMode']
export type TicketSort = components['schemas']['TicketSort']
export type Ticket = components['schemas']['Ticket']
export type Assignee = components['schemas']['Assignee']
export type AssigneeInput = components['schemas']['AssigneeInput']
export type TicketInput = components['schemas']['TicketInput']
export type Dashboard = components['schemas']['Dashboard']
export type ApiErrorBody = components['schemas']['ApiError']