# service-api/openapi.yaml
openapi: 3.1.0
info:
  title: API Codificar Desk
  version: 1.1.0
  description: Contrato HTTP do Codificar Desk para controle de chamados internos.
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Verifica a disponibilidade da API e do banco
      operationId: health
      responses:
        "200":
          description: Aplicação disponível
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: Banco de dados indisponível
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/dashboard:
    get:
      summary: Retorna os indicadores operacionais
      operationId: getDashboard
      responses:
        "200":
          description: Indicadores carregados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/assignees:
    get:
      summary: Lista os responsáveis
      operationId: listAssignees
      responses:
        "200":
          description: Responsáveis disponíveis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignee"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Adiciona um membro à equipe
      operationId: createAssignee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssigneeInput"
      responses:
        "201":
          description: Membro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignee"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/assignees/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      summary: Atualiza um membro da equipe
      operationId: updateAssignee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssigneeInput"
      responses:
        "200":
          description: Membro atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignee"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      summary: Remove um membro sem chamados associados
      operationId: deleteAssignee
      responses:
        "204":
          description: Membro removido
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/tickets:
    get:
      summary: Lista e filtra chamados
      operationId: listTickets
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/Priority"
        - name: assigneeId
          in: query
          schema:
            type: integer
            format: int64
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/TicketSort"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Chamados encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Cria um chamado
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Chamado criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tickets/{id}:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      summary: Retorna os detalhes de um chamado
      operationId: getTicket
      responses:
        "200":
          description: Chamado encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      summary: Atualiza um chamado
      operationId: updateTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "200":
          description: Chamado atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tickets/{id}/status:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    patch:
      summary: Altera diretamente o status de um chamado
      operationId: changeTicketStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusInput"
      responses:
        "200":
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    TicketID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Chamado não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ValidationError:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: A operação conflita com o estado atual
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: Erro interno
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    Priority:
      type: string
      enum: [low, medium, high]
    TicketStatus:
      type: string
      enum: [open, in_progress, resolved, closed]
    AssignmentMode:
      type: string
      enum: [manual, automatic]
    TicketSort:
      type: string
      enum: [priority_desc, opened_desc, opened_asc, updated_desc]
    Ticket:
      type: object
      required:
        - id
        - title
        - description
        - requesterName
        - priority
        - status
        - assigneeId
        - assigneeName
        - assignmentMode
        - openedAt
        - resolvedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        requesterName:
          type: string
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/TicketStatus"
        assigneeId:
          type: integer
          format: int64
        assigneeName:
          type: string
        assignmentMode:
          $ref: "#/components/schemas/AssignmentMode"
        openedAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketInput:
      type: object
      required:
        - title
        - description
        - requesterName
        - priority
        - status
        - assignmentMode
        - assigneeId
        - redistribute
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          minLength: 1
        requesterName:
          type: string
          minLength: 1
          maxLength: 120
        priority:
          $ref: "#/components/schemas/Priority"
        status:
          $ref: "#/components/schemas/TicketStatus"
        assignmentMode:
          $ref: "#/components/schemas/AssignmentMode"
        assigneeId:
          type: integer
          format: int64
        redistribute:
          type: boolean
    StatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/TicketStatus"
    Assignee:
      type: object
      required: [id, name, active, openTickets, completedTickets, lastAssignedAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        active:
          type: boolean
        openTickets:
          type: integer
        completedTickets:
          type: integer
        lastAssignedAt:
          type: [string, "null"]
          format: date-time
    AssigneeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
    Dashboard:
      type: object
      required: [total, open, inProgress, resolved, closed, assignees, nextAssignee]
      properties:
        total:
          type: integer
        open:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer
        closed:
          type: integer
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/Assignee"
        nextAssignee:
          oneOf:
            - $ref: "#/components/schemas/Assignee"
            - type: "null"
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
    Health:
      type: object
      required: [status, database]
      properties:
        status:
          type: string
        database:
          type: string